allow parameter of a collection to be used for image source
If a parameter is part of a collection, e.g. text parameter
"imglink" inside collection parameter "Client" then this parameter
"Client.imglink" can now be used for an image source. Previously
the parameter was not found when an image source was evaluated. This
also allows to use a parameter of a collection for a data source (once
list parameters inside collections are possible in the Designer).