Evaluate parameters inside list
Parameters inside a List-Parameter which contain an expression are
now evaluated for each row. Other parameters of the same row
can be accessed in the expression.

Further the internal row_number Parameter is initially set for
all list rows and not during iteration when used as a data source.

Fixes #12