Add test infrastructure and render tests based on demos
This is the initial approach to detect if the generated documents
"look" like they should. The comparison is done based on a checksum.
In order to reproducible generate documents all dynamic meta data
need to be set to a predefined value. For pdf and xlsx this is only
the creation date.
For reference the checksum is either a generated reported or a
file with a name ending with `.checksum` containing the checksum
for that test.
Data for test report generation is either read from data.json or
directly from the report definition.

See tests/README.md for additional details.

To run:
```
poetry run pytest
```
If you want the coverage data as well, run:
```
poetry run pytest --cov=reportbro
```
Test coverage is shown after the tests are run.

Other changes:

* add test to check if the watermark logo is part of the resources

Note: poetry > 1.2.2 is now required